Extracted the core functionality of Graph2D
Extracted the core functionality of Graph2D to Base2D to prepare for different graph types like bar graphs.
